+++ CHANGELOG.rst
@@ -26,7 +26,7 @@ Unreleased
 * Modified scripts/feature_toggle_report_generator to work based on envs
 * Removed confluence integration
 * Moved HtmlRenderer to its own file
-* Add ADR for new DjangoSettingToggle.
+* Add ADR for new SettingToggle. (see 0003-django-setting-toggles.rst)
 
 [NOTE: None of these versions have actually been released to PyPI, even though
 the version number has been bumped.]